Ope 1985 register process instance finished (#2037)
* impl(backend): moving packages

* impl(backend): fix imports

* impl(backend): moving register of event to the right place

* impl(backend): cleaning up

* impl(backend): cleaning up

* impl(backend): cleaning up

* impl(backend): Updating logic for getting metrics

* impl(backend): Clean up

* impl(backend): stabilizing import several versions test

* impl(backend): stabilizing import several versions test

* impl(backend): stabilizing import several versions test

* impl(backend): stabilizing import several versions test

* impl(backend): stabilizing import several versions test

* impl(backend): shortening ProcessStatisticsAfterArchivingIT

* impl(backend): updating to use bulk request

* impl(backend): updating how we set id and timestamp of event

* impl(backend): removing unused import

* impl(backend): updating tests

* impl(backend): moving back packages prior moved by mistake

* impl(backend): checking test logic

* impl(backend): reduce amount of data imported

* impl(backend): reduce timeout time

* impl(backend): first fail instances, then finish it

* impl(backend): flush data

+++ webapp/src/main/java/io/camunda/operate/webapp/management/dto/UsageMetricDTO.java
@@ -8,13 +8,13 @@ package io.camunda.operate.webapp.management.dto;
 import java.util.Objects;
 
 public class UsageMetricDTO {
-  private int total;
+  private long total;
 
-  public int getTotal() {
+  public long getTotal() {
     return total;
   }
 
-  public UsageMetricDTO setTotal(int total) {
+  public UsageMetricDTO setTotal(long total) {
     this.total = total;
     return this;
   }
